removed query at pkgName, moved convert functions into helper package to conslidate for both backends, updated isOccurrence schema
