+++ pkg/assembler/backends/neo4j/artifact.go
@@ -161,12 +161,3 @@ func generateModelArtifact(algorithm, digest string) *model.Artifact {
 	}
 	return &artifact
 }
-
-// TODO: maybe use generics for ArtifactInputSpec and ArtifactSpec?
-func convertArtInputSpecToArtSpec(artInput *model.ArtifactInputSpec) *model.ArtifactSpec {
-	artSpec := model.ArtifactSpec{
-		Algorithm: &artInput.Algorithm,
-		Digest:    &artInput.Digest,
-	}
-	return &artSpec
-}
