+++ pkg/assembler/backends/neo4j/ghsa.go
@@ -192,12 +192,3 @@ func generateModelGhsa(id string) *model.Ghsa {
 	}
 	return &ghsa
 }
-
-// TODO: maybe use generics for GHSAInputSpec and GHSASpec?
-func convertGhsaInputSpecToGhsaSpec(ghsaInput *model.GHSAInputSpec) *model.GHSASpec {
-	ghsaID := strings.ToLower(ghsaInput.GhsaID)
-	ghsaSpec := model.GHSASpec{
-		GhsaID: &ghsaID,
-	}
-	return &ghsaSpec
-}
