+++ pkg/assembler/backends/neo4j/pkg.go
@@ -790,42 +790,3 @@ func generateModelPackage(pkgType, namespaceStr, nameStr string, versionValue, s
 	}
 	return &pkg
 }
-
-// TODO: maybe use generics for PkgInputSpec and PkgSpec?
-func convertPkgInputSpecToPkgSpec(pkgInput *model.PkgInputSpec) *model.PkgSpec {
-	qualifiers := convertQualifierInputToQualifierSpec(pkgInput.Qualifiers)
-	matchEmpty := false
-	if len(qualifiers) == 0 {
-		matchEmpty = true
-	}
-	var version string = ""
-	if pkgInput.Version != nil {
-		version = *pkgInput.Version
-	}
-	var subpath string = ""
-	if pkgInput.Subpath != nil {
-		subpath = *pkgInput.Subpath
-	}
-	pkgSpec := model.PkgSpec{
-		Type:                     &pkgInput.Type,
-		Namespace:                pkgInput.Namespace,
-		Name:                     &pkgInput.Name,
-		Version:                  &version,
-		Subpath:                  &subpath,
-		Qualifiers:               qualifiers,
-		MatchOnlyEmptyQualifiers: &matchEmpty,
-	}
-	return &pkgSpec
-}
-
-func convertQualifierInputToQualifierSpec(qualifiers []*model.PackageQualifierInputSpec) []*model.PackageQualifierSpec {
-	pkgQualifiers := []*model.PackageQualifierSpec{}
-	for _, quali := range qualifiers {
-		pkgQualifier := &model.PackageQualifierSpec{
-			Key:   quali.Key,
-			Value: &quali.Value,
-		}
-		pkgQualifiers = append(pkgQualifiers, pkgQualifier)
-	}
-	return pkgQualifiers
-}
