+++ pkg/assembler/backends/neo4j/src.go
@@ -558,23 +558,3 @@ func generateModelSource(srcType, namespaceStr, nameStr string, commitValue, tag
 	}
 	return &src
 }
-
-// TODO: maybe use generics for SourceInputSpec and SourceSpec?
-func convertSrcInputSpecToSrcSpec(srcInput *model.SourceInputSpec) *model.SourceSpec {
-	var tag string = ""
-	if srcInput.Tag != nil {
-		tag = *srcInput.Tag
-	}
-	var commit string = ""
-	if srcInput.Commit != nil {
-		commit = *srcInput.Commit
-	}
-	srcSpec := model.SourceSpec{
-		Type:      &srcInput.Type,
-		Namespace: &srcInput.Namespace,
-		Name:      &srcInput.Name,
-		Tag:       &tag,
-		Commit:    &commit,
-	}
-	return &srcSpec
-}
